Let Google Translate detect language
Changed the standard language version of the script to not pass a
source language, so it can translate nearly any language into English
(or the language of your choice, if you chang that in the script)
automatically. Works with most languages I tried, though oddly not
Korean (Chinese and Thai, for instance, work fine though)

Also added a Large Type display to show the translated text on-screen
directly.